refactor: Remove unused clockify api
